+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ Changelog
 0.1.1 (upcoming)
 ----------------
 
-* ...
+* Fixed issue where most recent value was not used
+  in subsequent logaugment.add calls.
 
 0.1 (August 30, 2014)
 ---------------------